Sorted classes into sperate folders
