+++ phizapi/Models/Device/DeviceLoginRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace phizapi.Models.Device
+{
+    // DeviceLoginRequest is used to login a device with its mac-address
+    public class DeviceLoginRequest()
+    {
+        public string mac_address { get; set; }
+    }
+}
