+++ phizapi/Models/Image/ImagePersonUpdate.cs
@@ -1,0 +1,9 @@
+﻿namespace phizapi.Models.Image
+{
+    // ImagePersonUpdate is used to update the person_id of an image
+    public class ImagePersonUpdate
+    {
+        public string person_id { get; set; }
+
+    }
+}
