+++ phizapi/Models/Person/CreatePerson.cs
@@ -1,0 +1,9 @@
+﻿namespace phizapi.Models.Person
+{
+    // CreatePerson is used to create a new person in the database
+    public class CreatePerson
+    {
+        public string name { get; set; } = "";
+    }
+  
+}
