+++ phizapi/Models/Person/CustomDetails.cs
@@ -1,0 +1,10 @@
+﻿namespace phizapi.Models.Person
+{
+    // CustomDetails is used to store custom details about a person
+    public class CustomDetails
+    {
+        public string field_name { get; set; }
+        public string value { get; set; }
+    }
+
+}
