+++ phizapi/Models/Person/CustomDetailsRemove.cs
@@ -1,0 +1,7 @@
+﻿namespace phizapi.Models.Person
+{
+    public class CustomDetailsRemove
+    {
+        public string field_name { get; set; }
+    }
+}
