+++ phizapi/Models/User/CustomData.cs
@@ -1,0 +1,8 @@
+﻿namespace phizapi.Models.User
+{
+    public class CustomData
+    {
+        string key { get; set; } = "";
+        object value { get; set; }
+    }
+}
