+++ phizapi/Models/User/LoginRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace phizapi.Models.User
+{
+    public class LoginRequest
+    {
+        public string username { get; set; } = "";
+        public string password { get; set; } = "";
+    }
+
+}
