+++ phizapi/Models/User/Role.cs
@@ -1,0 +1,9 @@
+﻿namespace phizapi.Models.User
+{
+
+    public enum Role
+    {
+        Admin,
+        User,
+    }
+}
